+++ types/dexie-event-set.d.ts
