+++ types/errors.d.ts
