+++ types/promise-extended.d.ts
