+++ types/then-shortcut.d.ts
